% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayize.R
\name{arrayize}
\alias{arrayize}
\title{From a named vector with "BUGS format" names (v[1], m[2,3]), create a list of vectors and matrices}
\usage{
arrayize(x)
}
\description{
From a named vector with "BUGS format" names (v[1], m[2,3]), create a list of vectors and matrices
}
