\encoding{utf8}
\name{blocker}
\alias{blocker}
\title{Beta blockers to prevent mortality after myocardial infarction}
\description{A dataset of 22 trials investigating beta blockers versus control to prevent mortality after myocardial infarction. Number of events and sample size.}
\format{A meta-analysis dataset containing 44 rows of arm-based data (responders and sample size).}
\source{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a),
\emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials},
Medical Decision Making 33(5):607-617.
[\href{https://doi.org/10.1177/0272989X12458724}{doi:10.1177/0272989X12458724}]

J.B. Carlin (1992),
\emph{Meta-analysis for 2 × 2 tables: a Bayesian approach},
Statistics in Medicine 11(2):141-158.
[\href{https://doi.org/10.1002/sim.4780110202}{doi:10.1002/sim.4780110202}]
}
