\encoding{utf8}
\name{depression}
\alias{depression}
\title{Treatment response in major depression}
\description{A dataset of 111 trials investigating 12 treatments for major depression on treatment response. Treatment response was defined as a reduction of at least 50\% from the baseline score on the HAM-D or MADRS at week 8 (or, if not available, another time between week 6 and 12).}
\format{A network meta-analysis dataset containing 224 rows of arm-based data (responders and sample size).}
\source{
Cipriani et al. (2009),
\emph{Comparative efficacy and acceptability of 12 new-generation antidepressants: a multiple-treatments meta-analysis},
Lancet 373(9665):746-758.
[\href{https://doi.org/10.1016/S0140-6736(09)60046-5}{doi:10.1016/S0140-6736(09)60046-5}]
}
