\encoding{utf8}
\name{dietfat}
\alias{dietfat}
\title{Effects of low-fat diets on mortality}
\description{A dataset of 10 trials investigating low-fat diet versus control diet for mortality. Number of events and exposure in person-years.}
\format{A meta-analysis dataset containing 20 rows of arm-based data (responders, exposure, and sample size).}
\source{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a),
\emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials},
Medical Decision Making 33(5):607-617.
[\href{https://doi.org/10.1177/0272989X12458724}{doi:10.1177/0272989X12458724}]

Hooper et al. (2000),
\emph{Reduced or modified dietary fat for preventing cardiovascular disease},
Cochrane Database of Systematic Reviews 2:CD002137.
[\href{https://doi.org/10.1002/14651858.CD002137}{doi:10.1002/14651858.CD002137}]
}
