\encoding{utf8}
\name{smoking}
\alias{smoking}
\title{Psychological treatments to aid smoking cessation}
\description{A dataset of 24 trials investigating four psychological treatments and no treatment for smoking cessation. The outcome is the number of people who stopped smoking.}
\format{A network meta-analysis dataset containing 50 rows of arm-based data (responders and sample size).}
\source{
Lu and Ades (2006),
\emph{Assessing Evidence Inconsistency in Mixed Treatment Comparisons},
Journal of the American Statistical Society, 101(474):447-459.
[\href{https://doi.org/10.1198/016214505000001302}{doi:10.1198/016214505000001302}]

Hasselblad (1998),
\emph{Meta-analysis of multitreatment studies},
Medical Decision Making 18(1):37-43.
[\href{https://doi.org/10.1177/0272989X9801800110}{doi:10.1177/0272989X9801800110}]
}
