% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopIfNotConsistent.R
\name{stopIfNotConsistent}
\alias{stopIfNotConsistent}
\title{Stop if the results were not derived from an internally consistent model type}
\usage{
stopIfNotConsistent(result, method)
}
\description{
Stop if the results were not derived from an internally consistent model type
}
