\encoding{utf8}
\name{thrombolytic}
\alias{thrombolytic}
\title{Thrombolytic treatment after acute myocardial infarction}
\description{A dataset of 28 trials investigating eight thrombolytic treatments administered after a myocardial infarction. The outcome is mortality after 30-35 days.}
\format{A network meta-analysis dataset containing 58 rows of arm-based data (responders and sample size).}
\source{
Lu and Ades (2006),
\emph{Assessing Evidence Inconsistency in Mixed Treatment Comparisons},
Journal of the American Statistical Society, 101(474):447-459.
[\href{https://doi.org/10.1198/016214505000001302}{10.1198/016214505000001302}]

Boland et al. (2003),
\emph{Early thrombolysis for the treatment of acute myocardial infarction: a systematic review and economic evaluation},
Health Technology Assessment 7(15):1-136.
[\href{https://doi.org/10.3310/hta7150}{doi:10.3310/hta7150}]
}
