\encoding{utf8}
\name{parkinson}
\alias{parkinson}
\alias{parkinson_diff}
\alias{parkinson_shared}
\title{Mean off-time reduction in Parkinson's disease}
\description{A dataset of seven trials investigating four treatments and placebo for Parkinson's disease. The outcome is mean off-time reduction.}
\format{ \itemize{
\item{parkinson: }{A network meta-analysis dataset containing fifteen rows of arm-based data (mean, standard deviation, and sample size).}
\item{parkinson_diff: }{A network meta-analysis dataset containing fifteen rows of contrast-based data.}
\item{parkinson_shared: }{A network meta-analysis dataset containing mixed arm-based and contrast-based data.}
} }
\source{
Franchini et al. (2012),
\emph{Accounting for correlation in network meta-analysis with multi-arm trials},
Research Synthesis Methods, 3(2):142-160.
[\href{https://doi.org/10.1002/jrsm.1049}{doi:10.1002/jrsm.1049}]
}
